+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add `AZURE_KEY_VAULT_FILE` and `AZURE_KEY_VAULT_PASSWORD` environment variables in `ci`.
 ### Tests
 - Test `AzureKeyVaultConfig` with `AZURE_KEY_VAULT_FILE` environment variable.
+- Test `NuGetKeyVaultSignTool` in a package.
+- Test `AzureSignTool` in a assembly file.
 
 [vNext]: ../../compare/1.0.0...HEAD
 [1.0.0]: ../../compare/1.0.0